Test update required by dirt-protection in history change.
The protected submodule test has a side-effect brought about by the re-use of the same empty file for foo.txt, moo.txt and the submodule bar/comp.  Because the same file blob is used, foo.txt when removed from history was no longer removed because it was protected by bar/comp in master HEAD.

This re-writes the commits in that repository so that moo.txt and bar/comp are different blobs to foo.txt, so that foo.txt can be removed from the history and the test pass.